AS7-5116 add-user script not always checking password in interactive mode
